feat: Take category from current route and keep it in the store
Pretty complicated to get right: We want to derive the current category
from the URL when the user bookmarks it or presses F5. Also, we want to
show all posts (regardless of category) when the user is at '/'
BUT we also want to allow the user to change the categories by clicking
on the router links above. Current solution:
The router will pass the current category as props initially and on
route change. The CategoryView will pick these up and in turn trigger
the appropriate action to store the current category in the app state.
This happens either on the initial mount or on route change.