fix: Moved dispatch to the action, rather than the reducer. Doh!
After reading https://github.com/reactjs/redux/issues/368, I've come to
realize why reducers need to be pure and especially why they can't and
shouldn't dispatch actions while the state is in limbo. Kinda like the
same as react forbidding setState() in the render method.
So, moved the trigger for the fetch posts action to the change category
action creator.