chore: Kicking out yarn in favor of npm
After cleaning up my old installation, yarn won't let me reinstall it
the "simple" way (as cmd script in the node directory), but instead
wants me to install it via .msi or even add it to the global path.
Not good enough, sorry guys. So, back to NPM, which has lockfiles now
anyway.