feat: Streamlined category handling to allow for category.name
Categories are fetched first now. The actual CategoryView ist only
rendered once they become available. This is done to be able to select
the proper category name, which is not present when we first load the
page. Only the category path can be derived from the URL.
When the categories are available, we then proceed to map them by path,
so that the child components are able to display the proper category
name.