refactor: Use Post IDs instead of posts to decouple post component
We don't want to pass the list of posts to the category component,
because when one post changes, we will trigger a re-render of the whole
list of posts.
Instead, we keep track of the list of IDs and the post contents
separately. The Post component will only receive the ID and hence only
re-render when its post has changed.