feat: "AddPost" control with local state management
Post could well live in the global app state. But, in case of back and
forth (errors etc.), we would need to copy the global state over and
over again. Trying to keep this to the  component for now, also given
that we don't even want to save the post across the category views,
for various UX reasons.