feat: Implemented global feedback panel, show some API errors on it
The original intent had been to let the central API module post errors
to the feedback component directly. But since the API module is
not a React component, we can't connect it to the Redux store.
We *could* ex/import the store from the App module and use it directly,
but this is discouraged because it forces the store to be a singleton,
which makes testing harder, and also it would prevent server rendering
from working, as the store would be shared across all users requests
 (because, singleton).
The only method here would be to pass dispatch down to the API module.
But then I realized that this has been bugging me all along: User
feedback should not be the concern of the API, in fact the API should
not even know that it is called from a GUI or whatever.
So, I yanked it out and now every actions's API call is wrapped with
catch, which it needs to be anyway.